fix: Normalize whitespace in attribute values correctly
Previously, attribute values were normalized according to the rules for
non-CDATA attributes, but this was incorrect and based on a misreading
of the spec.

Attribute values are now correctly parsed as CDATA, meaning that
whitespace is not collapsed or trimmed and whitespace character entities
are resolved to their respective characters rather than being normalized
to spaces (which was incorrect even by the non-CDATA rules!).

See <https://www.w3.org/TR/2008/REC-xml-20081126/#AVNormalize>

Fixes #6