feat: Rewrite the parser!
While the previous parser was good, it relied heavily on complex regular
expressions. This helped keep it extremely small, but also left it open
to the possibility of regex denial of service bugs when parsing unusual
or maliciously crafted input.

The new parser uses a less interesting but overall safer approach, and
employs regular expressions only sparingly and in ways that aren't risky
(they're now only used as performance optimizations rather than as the
basis for the entire parser).

See `CHANGELOG.md` and `API.md` for more details on everything that's
changed. It's a lot!