Unskip some malformed UTF-8 conformance tests
Several conformance tests were being skipped because they contained
malformed byte sequences that couldn't be loaded as UTF-8. This was
awkward, because the point of these tests was to verify that parsing
fails correctly on malformed byte sequences in documents that are
supposed to be UTF-8, but since we couldn't even load the test files, we
couldn't pass them to the parser.

It turns out loading these tests as UTF-16LE rather than UTF-8 works
fine, so we can now unskip them and allow the parser to succeed at
failing to parse them. 🎉

Fixes #16