Improve parsing performance by as much as 2x
This removes the single most expensive parsing step (splitting the input
string into an array of multi-byte aware characters) by avoiding the
array creation entirely and being smarter about how we determine whether
multi-byte mode is necessary.

On some documents (particularly medium- to large-sized documents), this
appears to result in parsing times as much as twice as fast as the
previous approach in Node.js 16.17.0.