Add optional byte offsets to nodes
This adds a new parser option named `includeOffsets`, which defaults to
`false`. When `includeOffsets` is `true`, the byte offset of each node
in the input string will be made available via an `offset` property on
the node.

Note that this offset doesn't take into account any carriage return
(`\r`) characters in the input string because carriage returns are
removed during a normalization step before parsing begins.

Closes #24